refactor: Style Devcon claim page. Add hero section and og image (#30)
